requirements.txt
streamlit
gspread
google-auth
uuid
datetime
crewai
langchain-openai
langchain-community
ollama